Don't create salt, argon2id already does that
Also prepare to generate diceware usernames which would prevent
targeted attacks but remain easy to memorize. Being usernames, these
need not be crytographically secure. They need only be unique if they
are also used as a user ID, which may be the direction I go to avoid
having an additional UUID column.